abbyyR: R Client for the Abbyy Cloud OCR

OCR images, barcodes, forms, documents with machine readable zones, e.g. passports, right from R. Get the results in any one of a wide variety of formats, from text to XML. The package provides access to Abbyy Cloud OCR. For more information, see http://ocrsdk.com/. Details about results of calls to the API can be found at http://ocrsdk.com/documentation/specifications/status-codes/.

To learn how to use abbyyR, see this vignette: vignettes/Overview_of_abbyyR.html. Or, see how to scrape text from a folder of static Wisconsin Ads storyboards: vignettes/wiscads.html.

You need to get credentials (application ID and password) to use this application. If you haven't already, get these at http://ocrsdk.com/. And set these using setapp