Process Document

This function processes several images for the same task and results in a multi-page document. For instance, upload pages of the book individually via submitImage to the same task. And then process it via ProcessDocument to get a multi-page pdf.

processDocument(taskId = NULL, language = "English",
  profile = c("documentConversion", "documentArchiving", "textExtraction",
  "fieldLevelRecognition", "barcodeRecognition"), textType = c("normal",
  "typewriter", "matrix", "index", "ocrA", "ocrB", "e13b", "cmc7", "gothic"),
  imageSource = c("auto", "photo", "scanner"),
  correctOrientation = c("true", "false"), correctSkew = c("true", "false"),
  readBarcodes = c("false", "true"), exportFormat = c("txt",
  "txtUnstructured", "rtf", "docx", "xlsx", "pptx", "pdfSearchable",
  "pdfTextAndImages", "pdfa", "xml", "xmlForCorrectedImage", "alto"),
  description = NULL, pdfPassword = NULL, ...)

Arguments

taskId
Only tasks with Submitted, Completed or NotEnoughCredits status can be processed using this function.

language
String. Optional; default: English

profile
String. Optional; default: documentConversion Options: documentConversion, documentArchiving, textExtraction, fieldLevelRecognition, barcodeRecognition

textType
String. Optional; default: normal Options: normal, typewriter, matrix, index, ocrA,ocrB, e13b,cmc7, gothic

imageSource
String. Optional; default: auto Options: auto, photo, scanner

correctOrientation
String. Optional; default: true. Options: true or false

correctSkew
String. Optional; default: true. Options: true or false

readBarcodes
Optional; Options: true or false

exportFormat
optional, default: txt options: txt, txtUnstructured, rtf, docx, xlsx, pptx, pdfSearchable, pdfTextAndImages, pdfa, xml, xmlForCorrectedImage, alto

description
Optional; default: ""

pdfPassword
Optional; default: NULL

…
Additional arguments passed to abbyy_GET.

Value

data.frame with details of the task associated with the submitted Document

References

http://ocrsdk.com/documentation/apireference/processDocument/

Examples

# NOT RUN {
processDocument(taskId = "task_id")
# }