Process Remote Image

This function gets Information about a particular application

processRemoteImage(img_url = NULL, language = "English",
  profile = c("documentConversion", "documentArchiving",
  "textExtraction", "fieldLevelRecognition", "barcodeRecognition"),
  textType = c("normal", "typewriter", "matrix", "index", "ocrA", "ocrB",
  "e13b", "cmc7", "gothic"), imageSource = c("auto", "photo", "scanner"),
  correctOrientation = c("true", "false"), correctSkew = c("true",
  "false"), readBarcodes = c("false", "true"), exportFormat = c("txt",
  "txtUnstructured", "rtf", "docx", "xlsx", "pptx", "pdfSearchable",
  "pdfTextAndImages", "pdfa", "xml", "xmlForCorrectedImage", "alto"),
  description = NULL, pdfPassword = NULL, ...)

Arguments

img_url
Required; url to remote image

language
Optional; default: English

profile
String. Optional; default: documentConversion Options: documentConversion, documentArchiving, textExtraction, fieldLevelRecognition, barcodeRecognition

textType
String. Optional; default: normal. Specifies the type of the text on a page. Options: normal, typewriter, matrix, index, ocrA,ocrB, e13b,cmc7, gothic

imageSource
String. Optional; default: auto Options: auto, photo, scanner

correctOrientation
String. Optional; default: true. Options: true or false

correctSkew
String. Optional; default: true. Options: true or false

readBarcodes
Optional; default:

exportFormat
Optional; default: txt. Must be one of the following: txt, rtf, docx, xlsx, pptx, pdfSearchable, pdfTextAndImages, pdfa, xml, alto

description
Optional; default: ""

pdfPassword
Optional; default: NULL

…
Additional arguments passed to abbyy_GET.

Value

Data frame with details of the task associated with the submitted Remote Image

References

http://ocrsdk.com/documentation/apireference/processRemoteImage/

Examples

# NOT RUN {
processRemoteImage(img_url = "img_url")
# }