Submit Image
Source: R/submitImage.R

Adds image to the existing task or creates a new task for the uploaded image. The new task isn't processed till processDocument or processFields is called.

submitImage(file_path = "", taskId = "", pdfPassword = "", ...)

Arguments

file_path
Required; Path to the document

taskId
Optional; Assigns image to the task ID specified. If an empty string is passed, a new task is created.

pdfPassword
Optional; If the pdf is password protected, put the password here.

…
Additional arguments passed to abbyy_POST.

Value

Data frame with all the details of the submitted image: id (task id), registrationTime, statusChangeTime, status (Submitted, Queued, InProgress, Completed, ProcessingFailed, Deleted, NotEnoughCredits), filesCount (No. of files), credits

References

http://ocrsdk.com/documentation/apireference/submitImage/

Examples

# NOT RUN {
submitImage(file_path="/images/image1.png", taskId="task_id", pdfPassword="pdf_password")
# }