Sets Application ID and Password

Set Application ID and Password. Needed for interfacing with Abbyy FineReader Cloud OCR SDK. Run this before anything else.

setapp(appdetails = NULL, force = FALSE)

Arguments

appdetails

A vector of app_id, app_password. Get these from http://ocrsdk.com/. Set them before you use other functions.

force

Force change the app_id and app_password stored in the environment

Details

The function looks for AbbyyAppId and AbbyyAppPassword in the environment. If it doesn't find them and if we don't want to force change in them, it looks for arguments. And if no arguments are passed, it asks for user to input the values.

References

http://ocrsdk.com/documentation/apireference/getApplicationInfo/

Examples

## Not run: ------------------------------------
# setapp(c("app_id", "app_password"))
## ---------------------------------------------